requirements.txt
asgiref==3.2.10
autopep8==1.5.4
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.2
django-allauth==0.44.0
django-ckeditor==5.9.0
django-filter==2.4.0
django-js-asset==1.2.2
django-livereload-server==0.3.2
djangorestframework==3.12.2
idna==2.10
oauthlib==3.1.0
Pillow==8.0.1
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.0
python3-openid==3.2.0
pytz==2020.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
urllib3==1.26.2
